Stromsensor Funktion noch richtig implementiert

+++ PREN_pico/src/StromSensor.h
@@ -2,7 +2,7 @@
 #define STROMSENSOR__H_
 
 void StromSensorInit(void);
-uint16_t read_Sensor_12V(void);
-uint16_t read_Sensor_5V(void);
+int read_Sensor_12V(void);
+int read_Sensor_5V(void);
 
 #endif
